requirements.txt
# Core dependencies
python>=3.11,<3.13

# Data processing
pandas==2.1.3
numpy==1.24.3
pyarrow==14.0.1

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0

# Experiment tracking
mlflow==2.9.1
optuna==3.4.0

# Azure ML
azure-ai-ml==1.12.0
azure-identity==1.15.0
azure-storage-blob==12.19.0
azure-storage-filedatalake==12.14.0

# Data validation & quality
great-expectations==0.18.8
pandera==0.17.2
evidently==0.4.8

# Utilities
joblib==1.3.2
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Monitoring
opencensus-ext-azure==1.1.13
applicationinsights==0.11.10

# Development tools
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
ruff==0.1.6
black==23.11.0
mypy==1.7.1
pre-commit==3.5.0

# Jupyter
jupyter==1.0.0
jupyterlab==4.0.9
ipywidgets==8.1.1
nbformat==5.9.2

# API (for serving)
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
